+++ packages/cli-next/src/commands/submit.ts
@@ -58,7 +58,7 @@ export default class Submit extends AirtableCommand {
      */
     private _appTemporaryPath?: string;
 
-    static description = 'Submit app for review for listing in the the Airtable Marketplace';
+    static description = 'Submit extension for review for listing in the the Airtable Marketplace';
 
     static examples = [
         `$ block submit
@@ -86,7 +86,7 @@ export default class Submit extends AirtableCommand {
             this.error(appConfigResult.err);
         }
         const appConfig = appConfigResult.value;
-        debug('loaded app config at %s', sys.path.relative(workingdir, appConfigPath));
+        debug('loaded extension config at %s', sys.path.relative(workingdir, appConfigPath));
 
         const remoteConfigPath = await findRemoteConfigPathByNameAsync(
             sys,